Clean up run-once idempotent block
