--- src/main/java/com/greghaskins/spectrum/EmptyBlock.java
@@ -1,9 +1,0 @@
-package com.greghaskins.spectrum;
-
-import com.greghaskins.spectrum.Spectrum.Block;
-
-final class EmptyBlock implements Block {
-    @Override
-    public void run() {
-    }
-}
